requirements.txt
Flask>=0.9
SQLAlchemy>=0.7.9
Flask-WTF
flask-bcrypt
flask-babel
flask-login
flask-script
flask-sqlalchemy
jsmin
cssmin
flask-assets
python-memcached
flask-kvsession
requests
alembic
oauth2
requests_oauthlib
sniffer
pyinotify
unidecode
markdown2